requirements.txt
Flask-SQLAlchemy==2.5.1
Flask-RESTful==0.3.9
Flask-Migrate==3.1.0
neo4j==5.8.0
pymysql==1.0.2
python-dotenv==0.20.0  # Para manejar variables de entorno
flask>=2.0,<2.1
werkzeug==2.0.3
